+++ .gitignore
@@ -267,3 +267,5 @@ $RECYCLE.BIN/
 *.lnk
 
 # End of https://www.toptal.com/developers/gitignore/api/linux,macos,windows,visualstudiocode,node,react,dotenv,git
+
+build
