+++ src/Routes/Pages/Equipment/Components/EquipCreate.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+const EquipCreate = () => {
+
+    return (
+        <>
+        </>
+    );
+}
+
+export default EquipCreate;
