+++ src/Routes/Pages/Equipment/Components/index.js
@@ -1,5 +1,5 @@
 export { default as EquipHeader } from './EquipHeader';
 export { default as EquipSearch } from './EquipSearch';
-export { default as EquipMain } from './EquipMain';
 export { default as Equip } from './Equip';
 export { default as EquipItem } from './EquipItem';
+export { default as EquipCreate } from './EquipCreate';
